pyproject.toml
[project]
name = "game-tom-benchmark"
version = "0.1.0"
description = "ToM benchmark: run testcases with Hugging Face Transformers"
readme = "eval/README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 100
target-version = "py39"
